+++ game/song.cc
@@ -67,7 +67,7 @@ Song::Song(nlohmann::json const& song): dummyVocal(TrackName::LEAD_VOCAL), rando
 		instrumentTracks.insert(std::make_pair(TrackName::GUITAR, InstrumentTrack(TrackName::GUITAR)));
 	}
 	if (song.count("bpm") > 0) {
-		m_bpms.push_back(BPM(0, 0, song.at("bpm").get<double>()));
+		m_bpms.push_back(BPM(0, 0, song.at("bpm").get<float>()));
 	}
 	collateUpdate();
 }
